Integrate ReactiveUI.Fody and improve dialog management
Included ReactiveUI.Fody to streamline the reactive properties in the ViewModels and refactored code to leverage this. Improvements were made to the dialog management approach, allowing for cleaner and more flexible interactions. Additional changes to `OptionsViewModel` include the addition of a `CopyTo` method, which transfers properties from one instance to another, and removal of an overloaded constructor.